#849 Moved meta-model related classes to the pojo-bl module to remove the need for dependency on 'platform-annotation-processors' by TG-based applications.
Also minor refactoring to make classes less mutable + javadoc.
