--- platform-annotation-processors/src/main/java/ua/com/fielden/platform/processors/metamodel/exceptions/EntityMetaModelException.java
@@ -1,9 +1,0 @@
-package ua.com.fielden.platform.processors.metamodel.exceptions;
-
-public class EntityMetaModelException extends RuntimeException {
-
-    public EntityMetaModelException(String message) {
-        super(message);
-    }
-
-}
