--- platform-annotation-processors/src/main/java/ua/com/fielden/platform/processors/metamodel/models/PropertyMetaModel.java
@@ -1,21 +1,0 @@
-package ua.com.fielden.platform.processors.metamodel.models;
-
-import ua.com.fielden.platform.processors.metamodel.IConvertableToPath;
-
-public class PropertyMetaModel implements IConvertableToPath {
-    private String path;
-    
-    public PropertyMetaModel(String path) {
-        this.path = path;
-    }
-    
-    @Override
-    public String toPath() {
-        return this.path;
-    }
-    
-    @Override
-    public String toString() {
-        return toPath();
-    }
-}
