+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/processors/metamodel/models/EntityMetaModel.java
@@ -1,15 +1,25 @@
 package ua.com.fielden.platform.processors.metamodel.models;
 
+import static java.lang.String.format;
+
 import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.processors.metamodel.IConvertableToPath;
-import ua.com.fielden.platform.processors.metamodel.exceptions.EntityMetaModelException;
+import ua.com.fielden.platform.processors.metamodel. exceptions.EntityMetaModelException;
 
+/**
+ * The base class for all domain entity meta-models that get generated by {@code MetaModelProcessor}.
+ * 
+ * @author TG Team
+ *
+ */
 public abstract class EntityMetaModel implements IConvertableToPath {
     private final String path;
     
-    public EntityMetaModel(String path) {
-        if (path == null)
-            throw new EntityMetaModelException(String.format("%s constructor received null as an argument.", this.getClass().getSimpleName()));
+    public EntityMetaModel(final String path) {
+        if (path == null) {
+            throw new EntityMetaModelException(format("%s constructor received null as an argument.", this.getClass().getSimpleName()));
+        }
+
         this.path = path;
     }
     
@@ -18,8 +28,9 @@ public abstract class EntityMetaModel implements IConvertableToPath {
     }
     
     protected final String joinPath(String propName) {
-        if (!this.path.isEmpty())
-            return String.format("%s.%s", this.path, propName);
+        if (!this.path.isEmpty()) {
+            return format("%s.%s", this.path, propName);
+        }
 
         return propName;
     }
@@ -43,8 +54,9 @@ public abstract class EntityMetaModel implements IConvertableToPath {
      */
     @Override
     public final String toPath() {
-        if (this.path.isEmpty())
+        if (path.isEmpty()) {
             return "this";
+        }
 
         return this.path;
     }
@@ -59,4 +71,5 @@ public abstract class EntityMetaModel implements IConvertableToPath {
      * @return
      */
     public abstract Class<? extends AbstractEntity> getEntityClass();
+
 }
